Bug: DICOM incorrect maxPixelValue
DICOM incorrect maxPixelValue
For DICOM images using photometric interpretation MONOCHROME1,
bioformats needs to invert the pixel the data. For 16 bit data we had
been using the WINDOW_WIDTH property as a max pixel value, instead this
actually provides us with the pixel range and is not clipped to 0.

Code has been updated to try and read the WINDOW_CENTER value (the
central value in the range) in order to work out an appropriate scale.
If this is not available or not a sensible value then we will use the
full max pixel range.

OME-Users message:
http://lists.openmicroscopy.org.uk/pipermail/ome-users/2015-September/00
5629.html

QA: https://www.openmicroscopy.org/qa2/qa/feedback/11401/

For info on the WINDOW_WIDTH and WINDOW_CENTER properties:
https://www.dabsoft.ch/dicom/3/C.11.2.1.2/

For info on the photometric interpretations:
https://www.medicalconnections.co.uk/kb/Photometric_Interpretations